+++ glide/estimators/__init__.py
@@ -1,5 +1,6 @@
 from glide.estimators.asi import ASIMeanEstimator
 from glide.estimators.classical import ClassicalMeanEstimator
+from glide.estimators.cluster_classical import ClusterClassicalMeanEstimator
 from glide.estimators.ipw_classical import IPWClassicalMeanEstimator
 from glide.estimators.ipw_ptd import IPWPTDMeanEstimator
 from glide.estimators.ppi import PPIMeanEstimator
@@ -10,6 +11,7 @@ from glide.estimators.stratified_ptd import StratifiedPTDMeanEstimator
 
 __all__ = [
     "ClassicalMeanEstimator",
+    "ClusterClassicalMeanEstimator",
     "StratifiedClassicalMeanEstimator",
     "IPWClassicalMeanEstimator",
     "PPIMeanEstimator",